requirements.txt
annotated-types==0.7.0
attrs==25.3.0
aws-cdk-lib==2.192.0
aws-cdk.asset-awscli-v1==2.2.233
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-lambda-python-alpha==2.192.0a0
aws-cdk.cloud-assembly-schema==41.2.0
bleach==6.2.0
boto3==1.38.4
botocore==1.38.4
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
constructs==10.4.2
contourpy==1.3.2
cycler==0.12.1
dacite==1.9.2
fonttools==4.57.0
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsii==1.111.0
kaggle==1.7.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
multimethod==1.12
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
packaging==25.0
pandas==2.2.3
patsy==1.0.1
phik==0.12.4
pillow==11.2.1
protobuf==6.30.2
publication==0.0.3
puremagic==1.28
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.12.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.0.0
six==1.17.0
sklearn-compat==0.1.3
statsmodels==0.14.4
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
typeguard==2.13.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
visions==0.8.1
webencodings==0.5.1
wordcloud==1.9.4
xgboost==3.0.0
